requirements.txt
rich
numpy
pyyaml
matplotlib
scikit-image
opencv-python
plotly
trimesh
open3d
pymeshlab
bpy
huggingface_hub

einops
lpips
pytorch-msssim
jaxtyping
kornia
transformers

boto3
botocore
wandb
nvitop
videoio
tqdm
easydict
apted

